Fix issues with setting proper create/update/published dates on pages/blog post
The issue was that both the Pages and BlogPost handlers were adding a StorageFilter for the CommonVersionRecord. This was already done by the CommonAspectHandler, so we ended up having 3 StorageFilter for pages and blog post content items. This lead to incorrect behavior when creating instances (the CommonVersionRecord was re-initialized 3 times). The fix is to remove the StorageFilter from BlogPost and Page. Also added a check in StorageFilter to throw an exception is this situation is be detected in the future.

--HG--
extra : convert_revision : svn%3A5ff7c347-ad56-4c35-b696-ccb81de16e03/trunk%4045936