Reversing order of IEnumerable<TService> resolved from IoC
Components are registered in dependency/priority order and
this is also the order events and hooks should fire. The default
order of Autofac is opposite what we need for enumerable services.

The behavior of resolving a single TService is unchanged - the
last-registered of TService is the only returned.

--HG--
branch : 1.x
extra : transplant_source : %87%22%FC%84%A1%0FP%3C%20w%B4%1EQ%CC%90%20%97%C1%EC%BF