Adding new ProbingExtensionLoader, and fixing DynamicCompilation priority.
The DyanmicExtensionLoader could randomly take over PrecompiledExtensionLoader
as the Priority was not set on ProbeExtensionEntry.

Also a new ProbingExtensionLoader is added and is used when DyanmicExtensionLoader
is disabled. It will compile an extension dynamically but cache the assembly in
the probing folder (Dependencies) so that it's not recompiled on each app start,
and such that there is no need to inspect dependency files.

--HG--
branch : 1.x