Preventing ProbingExtensionLoader from copying the asp.net temporarty assembly
If loaded by name in the PrecompiledLoader then the Fusion engine threw an
exception as the manifest name would not match the assembly name.
Actually the ASP.NET temp folder is cached using the hash of the compilation
unit which is the .csproj in our case, so it won't be rebuilt accross restarts,
unless necessary.

--HG--
branch : 1.x