Enabling the ENTER key to quickly navigate to the content type or part if results were found based on the filter.
If no results were found and hitting ENTER, the input is used as a suggested name for the content type or part to create.
If no results were found and clicking the primary button, the input is used as a suggested name for the content type or part to create.