Lock ShellDescriptor cache
The cache.dat files contains the enabled features for every tenants.
This file is accessed in read/write by all tenants when they are loaded
or changed. This can happen concurrently if tenants are activated
in parallel, if multiple servers are using the same file system, or if
multiple tenants get their features changed at the same time. This
commit doesn't fix the case of multiple web roles accessing the same
file system (WAWS).