Improving route based constraints
The IArchiveConstraint is a singleton as the routes are singletons too.
However the path resolution is per tenant, and when accessing archive
pages where lots of links are constructed simultaneously there are
collisions and DataReader/Connections failures. The solution is to use
the IAliasHolder which is also a singleton and doesn't make any query.