Changed argument names to match the way they are interpreted by the default implementation.
The term "skip" is generally interpreted as representing the number of records to skip.
However, the SearchService implementation interprets this value as the start page, not the number of records to skip.
We could have changed the implementation to match the interface, but doing it the other way around is safer.