Working around an model state / HTML helper issue.
The issue occurs when rendering an Html.CheckBox whose state contains a "true,false" string for example.
That string cannot be parsed into a boolean, so an error occurs.
This fix simply ignores any additional values for a given key (which presumably is what the default behavior is anyway, since the issue does not occur with the default value providers).