Layouts: Refactoring the way data is transferred between the layout editor and modal element editor dialog.
This change makes implementing element editors more intuitive, since they no longer rely on the initial postback from layout editor to the element editor dialog.
The ObjectStore is a simple database table that serves as a robust data transfer mechanism to share element state without the need for the host (content) to be persisted.
We could have used SessionState, but that would have added configuration requirements when hosting in a web farm.