Fixed site settings recipe step.
This fixes an issue where the remainder of the attributes on an XML element are not being processed if a certain attribute does not exist as a property on the part.
For example, the SiteCultures attribute on the element being imported would not match to a property on the SiteSettingsPart, so the remainder of the attributes would be skipped.