Fixed an issue where unchecked checkboxes would remain checked.
The issue was that by combining existing element data with posted element data, the existing items would retain their values in case no such item was posted, which is the case for unchecked checkboxes; their name/value pair will  be not posted, so the original value would be retained.
This is fixed by no longer combining existing data with hosted data, but instead relying that an element editor will post all data that it is concerned with.