Fixed binding issue for Enumeration field when using RadioList.
This fixes an issue when the Enumeration content field is configured to use radio buttons - in which case the editor shape will use the "Value" property instead of the "SelectedValues" property. Since the binding would only set the value through "SelectedValues", the stored value would be surrounded with a comma, which would not match with any of the enumeration values.