Fix Layout element localization
The ElementFactory did not always initialize the localizer of the
Element objects which caused the elements to end up with the
NullLocalizer.

This solution still has a major issue - the context of the localized
strings is wrong. It should be the source file in which the element
is defined, not "Orchard.Layouts.Services.ElementFactory".