Update ImageProfileManager.cs
Image cache is empty after every reboot of the Orchard app. It makey Dynamic ImageProfiles reprocess images after every reboot. It has a big impact on performance because all the images that use dynamic profiles are reprocessed when they are requested after a reboot. When you open a page with a number of images disks usage increases drammatically and the apps answer very slow to other requests

This is related with this issue https://github.com/OrchardCMS/Orchard/issues/5061