#5467: Fixes the issue where the "Set as homepage" setting is not retained when saving,
The issue was that when you save a content item (without publishing it) while checking the "Set as homepage" checkbox, it would lose its checked status after save.
With this change, the checkbox is again rendered but in a checked status, so that when you do publish, it will become the homepage.

Fixes #5467