Fixed that default layout data would not be applied.
This fixes the issue where the default layout data string would not be applied when creating new content items of a content type with the LayoutPart attached.
For example, creating new Form content items would start out with an empty canvas instead of with a default Form element and a submit button.