Falling back to a default base URL in case none is configured.
This fixes an issue where you run Setup from a command line and the recipe does not set BaseUrl on site settings.
The issue is that if the HttpContext does not have a URL, certain properties on the Request object will throw when constructing a URI with a null value.