Fixing Redis connection creation
With the previous code, when multiple threads tried to create
a Redis connection, the ConcurrentDictionary would have
called the factory delegate multiple times and added only one
of the resulting connections. This is by design, and a solution
is to use a Lazy object to defer the creation of the connection.
Multiple Lazy objects will be created but it doesn't have any impact.