Improved recipe import/reset execution by pausing sweep generator.
This prevents active background tasks from accessing tables that may no longer be there (in the case of a full site reset) by pausing the sweep generator.
To ensure that the sweep generator is resumed after the recipe has been executed, a new ActivateSweepGeneratorStep is enqueued at the end of the recipe execution.