Fixed a potential "ValueFactory attempted to access the Value property of this instance" issue.
The underlying cause of such an error is a recursive call issue, where some code in a background task accesses the current http context (which is a stand in), while its construction also (indirectly) accesses the http context in order to get the site content item and query the configured BaseUrl. In theory we could access this information straight from the database without loading the Site content item, but since the URL value doesn't really matter it is better to keep it simple.