[Fixes #5820] Fixing default layout support
The default layout was just set in the text editor, and would not be
taken into account if the Part settings were not saved. A user would
then think the default template was there but actually not.

The change here is to leave the text editor empty, and if the setting is
empty load the default from the service. To actually have an empty template
a user needs to provide an empty object.