Fixes Issue #5885
This should fix issue #5885, the DisplayAlias not having length validated. I arbitrarily set it to 1850 for the simple fact that, while the database can hold a 2048 character DisplayAlias, a URL should not be longer than ~2000 characters (per http://stackoverflow.com/questions/417142/what-is-the-maximum-length-of-a-url-in-different-browsers ). If the DisplayAlias is 2048, it's already over the limit and that's not the entire URL for the page, this gives some buffer room for domains.