Fixed order of exported content part elements.
This fixes an issue where the LayoutPart XML element would appear last in the content item XML element due to the fact that the <LayoutPart> element was being added only in the Exported event and not in the Exporting event.

To fix this, the <LayoutPart> element is no first added during the Exporting event, and later updated in the Exporting event.