Removed Orchard.Setup.Services feature.
This fixes a potential issue where modules would take on a dependency on Orchard.ImportExport (which in turn takes on a dependency on Orchard.Setup.Services) by removing the dependency of Orchard.Setup on Orchard.Setup.Services.
As described in ticket, this requires Orchard.Import to "manually" instantiate the SetupService, since it cannot rely on DI as the Orchard.Setup feature will be disabled after setup is done.

Fixes #5980