Fixed media library picker field script.
This fixes an issue where the code selecting the 'Save' button would match any button with the 'button.save' class.
The issue became apparent when working with Image element on a layout. The Element Editor Dialog also has a 'Save' button with the 'button save' class, en became hidden when no image was selected.