Improved class attribute logic.
This change fixes that the `"class"` attribute would be added to the tag even if there are no css classes to assign. This in turn has the effect that the root `Canvas` element is not rendering its outer tags needlessly by default (only if it has at least one attribute).