Reverted 42e71fe9fa2cc5337eb9239d68a08a2b680f3df2
Unfortunately this is breaking the setup. For some reason, when executed in a regular HTTP context,  CallContext  "loses" its data at the end of the request, causing recipe execution to fail. Until we understand why this is and how to best deal with it, I am reverting this file so that it at least works.