Fixed Dynamic Forms Bindings issue. Fixes #6470
ModelBinding failed due to the "null" key being specified. Instead, the type name of the model should be specified as the key (which happens by default when no key is specified). Not sure how this worked before, since this code has been there from the beginning, and file histories don;t show an obvious change. Perhaps due to the upgrade to newer ASP.NET MVC?