Work mostly on Alias admin views to make it a guideline.
Using Bootstrap modal instead of using javascript confirm()
Adding bulk action button instead of dropdownlist + button.
Modifications to the Pager
Removing the number of items displayed / page in the header with the filter options
Modifications to the AdminController to display a Notifier with the name of the contents in the same Notifier box when we do bulk actions "delete".