Redo "Improved performance of Count Hql queries"
This PR redos the performance improvement reverted in PR this https://github.com/OrchardCMS/Orchard/pull/7563
An alternative way of improving performance of Count Hql query compatible with SQLCE.
I have had to use a nested sql to avoid SQLCE limitation. Problem is those are not supported by Nibernate HQL so I translated main HQLquery  to SQL and then used it as nested sql of the general sql doing a count. Problem with this query is cannot be cached by NHibernate second level cache, so I removed SetCacheable line. Now performance is better than with reverted code, so I would prefer to use this version than previous one despite now the query is not cached by NHibernate. I submit this PR just in case it is interesting for you merging it. @sebastienros @paynecrl97