Log error when Recipient emailMessage.Recipients is null
When emailMessage.Recipients is null throws so the error is being logged by Orchard.Exceptions instead by the logger in this condition.
Repro steps:
Create a workflow that sends emails when a custom type has been created. 
Attach an email activity when this type has been created and dont fill it.