Improved performance of queries adding indexes
I have added indexes on creation migration, when db is empty cause for existing sites theses créate index opertations could lead to timeout exceptions when there is existing data to be indexed as happened to us in a 30 multitennt environment with a number of queries per tenant. So, it would be a task for the site administrator to add those indexes directly to the db by hand.