TaxonomyField was running Update when it should not
The UpdateEditorShape was doing the wrong checks to see whether it should process stuff. Basically, it was running when the ContentItem that contains the field does not have a LocalizationPart. That did not seem to break anything, really, but lead to weird validation messages, because they notifications from line 125 would pop up in the validation summary