Tweak TermsPartHandler to skip unpublished terms
When a term gets unpublished, rather than deleted, TermsPartHandler will spit out a  "given key was not present in the dictionary" error any time a content item that used to be linked requests its terms. Unpublishing terms is not supported out of the box but there are legitimate use cases for it.