Remove newly added terms before processing
`// Remove the newly added terms because they will get processed by the Published-Event
                termList.RemoveAll(t => t.Term.Id == term.Id);`

This line on 284 of the TaxonomyService claims to remove newly added terms, because they will be processed by their own published event anyway. I don't think it's currently doing that however, since it's comparing the id of the TermPart to the id of the TermContentItem record. This is unlikely to cause any issues besides some overhead, but there is a theoretical edge case where a TermContentItem could have the same id as an existing term, causing that term not to be processed at all.

I believe this is what was intended:
`termList.RemoveAll(t => t.Term.TermRecord.Id == term.Id);`